Keep encoder context with ColumnEncoder
