+++ Common/columnencoder.h
@@ -63,6 +63,7 @@ public:
 	
 	static	colVec				columnNames();
 	static	colVec				columnNamesEncoded();
+	static	const char*			extraOptionsPrefix()										{ return "JaspExtraOptions_"; }
 
 			bool				shouldEncode(const std::string & in);
 			bool				shouldDecode(const std::string & in);
